# light-controller-ui/api-docs.json
{"openapi":"3.0.1","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"paths":{"/signal-groups/{id}":{"get":{"tags":["signal-group-controller"],"operationId":"getSignalGroup","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/SignalGroupDTO"}}}}}},"put":{"tags":["signal-group-controller"],"operationId":"updateSignalGroup","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SignalGroupDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/SignalGroupDTO"}}}}}},"delete":{"tags":["signal-group-controller"],"operationId":"deleteSignalGroup","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK"}}}},"/phase/{id}":{"get":{"tags":["phase-controller"],"operationId":"getPhase","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PhaseDTO"}}}}}},"put":{"tags":["phase-controller"],"operationId":"updatePhase","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PhaseCreateRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PhaseDTO"}}}}}},"delete":{"tags":["phase-controller"],"operationId":"deletePhase","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK"}}}},"/intersections/{id}":{"get":{"tags":["intersection-controller"],"operationId":"getIntersection","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/IntersectionDTO"}}}}}},"put":{"tags":["intersection-controller"],"operationId":"updateIntersection","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/IntersectionUpdateRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/IntersectionDTO"}}}}}},"delete":{"tags":["intersection-controller"],"operationId":"deleteIntersection","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK"}}}},"/signal-groups":{"get":{"tags":["signal-group-controller"],"operationId":"getSignalGroups","parameters":[{"name":"intersectionId","in":"query","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/SignalGroupDTO"}}}}}}},"post":{"tags":["signal-group-controller"],"operationId":"createSignalGroup","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SignalGroupCreateRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/SignalGroupDTO"}}}}}}},"/phase":{"get":{"tags":["phase-controller"],"operationId":"getPhases","parameters":[{"name":"intersectionId","in":"query","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/PhaseDTO"}}}}}}},"post":{"tags":["phase-controller"],"operationId":"createPhase","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PhaseCreateRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PhaseDTO"}}}}}}},"/intersections":{"post":{"tags":["intersection-controller"],"operationId":"createIntersection","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/IntersectionCreateRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/IntersectionDTO"}}}}}}},"/intersections/list":{"get":{"tags":["intersection-controller"],"operationId":"getIntersectionList","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/IntersectionDTO"}}}}}}}}},"components":{"schemas":{"SignalGroupDTO":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"intersectionId":{"type":"integer","format":"int32"},"name":{"type":"string"},"currentColor":{"type":"string","enum":["GREEN","YELLOW","RED"]}}},"PhaseCreateRequest":{"type":"object","properties":{"intersectionId":{"type":"integer","format":"int32"},"sequence":{"type":"integer","format":"int32"},"greenDuration":{"type":"integer","format":"int32"},"yellowDuration":{"type":"integer","format":"int32"},"signalGroupIds":{"type":"array","items":{"type":"integer","format":"int32"}}}},"PhaseDTO":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"intersectionId":{"type":"integer","format":"int32"},"sequence":{"type":"integer","format":"int32"},"greenDuration":{"type":"integer","format":"int32"},"yellowDuration":{"type":"integer","format":"int32"},"signalGroupIds":{"type":"array","items":{"type":"integer","format":"int32"}}}},"IntersectionUpdateRequest":{"type":"object","properties":{"name":{"type":"string"},"active":{"type":"boolean"}}},"IntersectionDTO":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"name":{"type":"string"},"active":{"type":"boolean"},"lastTransitionTime":{"type":"string","format":"date-time"},"currentPhaseIndex":{"type":"integer","format":"int32"},"phases":{"type":"array","items":{"$ref":"#/components/schemas/PhaseDTO"}},"signalGroups":{"type":"array","items":{"$ref":"#/components/schemas/SignalGroupDTO"}}}},"SignalGroupCreateRequest":{"type":"object","properties":{"intersectionId":{"type":"integer","format":"int32"},"name":{"type":"string"},"currentColor":{"type":"string","enum":["GREEN","YELLOW","RED"]}}},"IntersectionCreateRequest":{"type":"object","properties":{"name":{"type":"string"}}}}}}